requirements.txt
# Utility
pydantic==1.10
ipython>=8.12
psutil>=5.9
pyyaml>=6.0
typing_extensions==4.5.0
loguru>=0.7.0
tqdm
python-dotenv

# Visualization
gradio
streamlit

tqdm>=4.65
matplotlib>=3.7
seaborn>=0.12

# Data handling
numpy==1.26
pandas>=2.0

# Image
opencv-python>=4.8
pillow>=9.5

# Others
python-magic==0.4.27
emoji>=2.6.0

# search
rank_bm25
voyager
duckdb

# NLP
mecab-python3
unidic-lite
fugashi
sentencepiece
sentence-transformers>=3.0

# Notion
notion-client
notion2md